Second working version: Transcript, unit tests, Makefile, jshint
There are several sets of changes in this commit:

Transcript: If gremlin-repl is run with stdout redirected, then
both the input and output of the REPL are written to stdout.
This produces a nice transcript of a session.

Unit Tests: First set of unit tests, currently just for TeeTransform,
a Transform stream that is used to create transcripts.

Makefile: Add a makefile that to run jshint and mocha unit tests.

jshint: Fixed various jshint warnings.