Various changes while working on gremlindocs methods tests. Most notably: Promises.
This commit started as an attempt to implement test gremlindocs-methods test.
But to test these methods it became clear that I needed to take advantage of promises.
This spurred various changes, which in an ideal world would have been rebased into
a more coherent set of commits, but I'm simply squashing all of the changes into
this commit and moving on.